backend/requirements.txt
asgiref==3.5.0
astroid==2.9.3
attrs==21.4.0
Django==4.0.3
django-admin==2.0.1
django-cors-headers==3.11.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-rest-auth==0.9.5
django-six==1.0.4
djangorestframework==3.13.1
environ==1.0
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.5.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
python-dotenv==0.19.2
python-environ==0.4.54
pytz==2021.3
screen==1.0.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
wrapt==1.13.3
xlwt==1.3.0
